requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

# PyTorch stack - Versiones más recientes con CUDA 12.1
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchtext==0.17.0
torchvision==0.17.0+cu121
triton==2.2.0

# TensorFlow stack - Versiones más recientes
tensorflow==2.16.1
tensorflow-estimator==2.16.1
tensorboard==2.16.1

# Compatibles con Python 3.10+ y numpy>=1.26
numpy==1.26.4
typing-extensions==4.11.0

# Vision / AI tools - Versiones actualizadas
onnx==1.16.0
onnxruntime-gpu==1.17.0
opencv-python==4.9.0.80
gfpgan==1.3.8
basicsr==1.4.2
facexlib==0.3.0
insightface==0.7.3
filterpy==1.4.5
opennsfw2==0.10.2

# UI
customtkinter==5.2.2
darkdetect==0.8.0
tkinterdnd2==0.3.0
tk==0.1.0

# Otros - Versiones actualizadas
coloredlogs==15.0.1
humanfriendly==10.0
pillow==10.2.0
sqlalchemy==2.0.31
addict==2.4.0
pydantic==2.8.0
pydantic-core==2.20.0
pandas-stubs==2.0.3.230814
lmdb==1.5.1
tqdm==4.66.1

# Dependencias adicionales para optimización GPU
nvidia-ml-py3==12.535.133
psutil==5.9.8
